requirements.txt
# FastAPI y dependencias
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0
starlette>=0.14.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# Utilidades
python-json-logger>=2.0.0
httpx>=0.24.0
tenacity>=8.0.0
aiohttp>=3.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0